some casting, type changes, etc. to reduce picky compiler warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@3821 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Features.h
@@ -72,7 +72,7 @@ public:
 	vtString m_name;
 
 	Array<bool> m_bool;
-	Array<int> m_short;
+	Array<short> m_short;
 	Array<int> m_int;
 	Array<float> m_float;
 	Array<double> m_double;
